LunPack Early Beta Release!
This is a 1.8 Exclusive! Uses forge as an environment and has a large
adventure side to it, with a touch of magic and industrial automation.
There is a server side and a client side. Please download the
appropriate sides for your environment (server if you are running a
server, client if you plan on playing this pack on your own personal
computer...). Last but not least, enjoy! P.S. There will be major
amounts of bugs due to most of the mods are in alpha and early beta
states, but otherwise everything should be stable.